[Update] change example code to single file; debug for review
