+++ plain-java-jdbc/Makefile
@@ -15,7 +15,7 @@
 .PHONY: all mysql build run
 
 all:
-	make mysql build run
+	make build run
 
 mysql:
 	mysql --host 127.0.0.1 --port 4000 -u root<src/main/resources/dbinit.sql
