--- plain-java-jdbc/src/main/java/com/pingcap/App.java
@@ -1,84 +1,0 @@
-// Copyright 2022 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.pingcap;
-
-import com.mysql.cj.jdbc.MysqlDataSource;
-import java.util.*;
-
-/**
- * Main class for the basic JDBC example.
- **/
-public class App 
-{
-    public static void main(String[] args) {
-        // 1. Configure the example database connection.
-
-        // 1.1 Create a mysql data source instance
-        MysqlDataSource mysqlDataSource = new MysqlDataSource();
-
-        // 1.2 Set server name, port, database name , username and password
-        mysqlDataSource.setServerName("localhost");
-        mysqlDataSource.setPortNumber(4000);
-        mysqlDataSource.setDatabaseName("test");
-        mysqlDataSource.setUser("root");
-        mysqlDataSource.setPassword("");
-
-        // Or you can use jdbc string instead
-        // mysqlDataSource.setURL("jdbc:mysql://{host}:{port}/game?user={user}&password={password}");
-
-        // 2. And then, create DAO to manager your data
-        PlayerDAO dao = new PlayerDAO(mysqlDataSource);
-
-        // 3. Run some simple example
-
-        // Create a player
-        dao.createPlayers(Collections.singletonList(new PlayerBean("test", 1, 1)));
-
-        // Get a player
-        PlayerBean testPlayer = dao.getPlayer("test");
-        System.out.printf("PlayerDAO.getPlayer:\n    => id: %s\n    => coins: %s\n    => goods: %s\n",
-                testPlayer.getId(), testPlayer.getCoins(), testPlayer.getGoods());
-
-        // Create players with bulk inserts
-        int addedCount = dao.bulkInsertRandomPlayers(1919, 114);
-        System.out.printf("PlayerDAO.bulkInsertRandomPlayers:\n    => %d total inserted players\n", addedCount);
-
-        // Count players amount
-        int count = dao.countPlayers();
-        System.out.printf("PlayerDAO.countPlayers:\n    => %d total players\n", count);
-
-        // Print 3 players
-        dao.printPlayers(3);
-
-        // 4. Getting further
-        PlayerBean player1 = new PlayerBean("1", 100, 0);
-        PlayerBean player2 = new PlayerBean("2", 114514, 20);
-
-        // Insert two players "by hand", using INSERT on the backend.
-        addedCount = dao.createPlayers(Arrays.asList(player1, player2));
-        System.out.printf("PlayerDAO.createPlayers:\n    => %d total inserted players\n", addedCount);
-
-        // Player 1 want to buy 10 goods from player 2
-        // It will cost 500 coins, but player 1 can't afford it
-        System.out.println("\nPlayerDAO.buyGoods:\n    => this trade will fail");
-        int updatedCount = dao.buyGoods(player2.getId(), player1.getId(), 10, 500);
-        System.out.printf("PlayerDAO.buyGoods:\n    => %d total update players\n", updatedCount);
-
-        // So player 1 have to reduce his incoming quantity to two
-        System.out.println("\nPlayerDAO.buyGoods:\n    => this trade will success");
-        updatedCount = dao.buyGoods(player2.getId(), player1.getId(), 2, 100);
-        System.out.printf("PlayerDAO.buyGoods:\n    => %d total update players\n", updatedCount);
-    }
-}
